Dict storage managment
From now on, dict that store data in *_data.py are raw (dict with str and no custom class) and then proceeded in __init__.py using from_dict() to change the raw dict into a dict with custom class. This change is needed to supress circular import errors and also making a better storage managment.